JSC cli should have a short option for useConcurrentJIT=0 https://bugs.webkit.org/show_bug.cgi?id=273528 rdar://127327863
Reviewed by Yusuke Suzuki.

This is a commonly used option so it should have a short option. This patch makes the
option `-s` for "synchronous compilation". Currently, this is used for signal handling
but that's only used by the test harness so we can move that to a long option, `--signal-expected`.

* Source/JavaScriptCore/jsc.cpp:
(printUsageStatement):
(CommandLine::parseArguments):
* Tools/Scripts/run-jsc-stress-tests:

Canonical link: https://commits.webkit.org/278231@main