RemoteImageBufferProxy should not check backend size https://bugs.webkit.org/show_bug.cgi?id=273623 rdar://problem/127423860
Remove the backend size check. GPUP will decide which backend will be
given, and as such the check cannot be made.

This is work towards removing the create<Backend> template parameter,
which is in turn blocking further work being able to construct new
RemoteImageBufferProxy instances out of new serialized image buffers.

* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::adjustImageBufferOptionsForTesting):
* Source/WebCore/platform/graphics/ImageBuffer.h:
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.cpp:
(WebKit::RemoteImageBufferProxy::didCreateBackend):
Avoid using canMapBackingStore(), as that is property of the backing store.
If the ImageBuffer has not received a backing store, this would reflect
the case what the create() caller expected to ask. In this case, it would
be same as the shouldMapAcceleratedImageBufferBackend().

* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.h:
(WebKit::RemoteImageBufferProxy::create):
* Source/WebKit/WebProcess/GPU/graphics/RemoteRenderingBackendProxy.cpp:
(WebKit::RemoteRenderingBackendProxy::shouldMapAcceleratedImageBufferBackend):
(WebKit::RemoteRenderingBackendProxy::createImageBuffer):
* Source/WebKit/WebProcess/GPU/graphics/RemoteRenderingBackendProxy.h: