Write data before creating mapping in mapToFile https://bugs.webkit.org/show_bug.cgi?id=273670 rdar://127472470
Reviewed by NOBODY (OOPS!).

Currently FileSystem::mapToFile creates a mapping, memcpys data into it, and then msyncs the data
back to the file. The issue I see with this in profiles is that we are paging in data from the file
system at page fault time when doing the memcpy only to instantly discard that data by overwriting
it. We should probably just write the data out to the file and then create a mapping instead.

I did think about using writev instead of write here, but technically the semantics of
`dispatch_data_apply` don't allow that without changing the interface of `mapToFile`, which probably
isn't worth it.

* Source/WTF/wtf/FileSystem.cpp:
(WTF::FileSystemImpl::mapToFile):