[visionOS] Refine transition for Image Fullscreen https://bugs.webkit.org/show_bug.cgi?id=272829 <rdar://126218680>
Reviewed by Aditya Keerthi.

Special case the transition to fullscreen when using Quick Look, since
scene swapping will take care of the animation, and we don't display the
fullscreen window.
Bump the QOS of the WorkQueue. It's only used as part of the fullscreen
transition, in reaction to user action.

* Source/WebKit/UIProcess/WebFullScreenManagerProxy.cpp:
(WebKit::sharedQuickLookFileQueue):
QOS bump.

* Source/WebKit/UIProcess/ios/fullscreen/WKFullScreenWindowControllerIOS.mm:
(-[WKFullScreenWindowController enterFullScreen:]):
Keep track of whether or not we're using Image Fullscreen. It impacts
the exit transition, and we don't want to depend on the exact moment the
image buffer is released in WebFullScreenManagerProxy.
(-[WKFullScreenWindowController _configureSpatialFullScreenTransition]):
Skip the spatial fullscreen transition window manipulation when using
Image Fullscreen.
(-[WKFullScreenWindowController _performSpatialFullScreenTransition:completionHandler:]):
Add a new Image Fullscreen code path for the transition in and out of
fullscreen. Hide the fullscreen window since we're not using it.
Remove the old branches at different points during the transition.
(-[WKFullScreenWindowController showUI]):
Remove the Image Fullscreen specific chrome management, the WebKit
window will be hidden.

Canonical link: https://commits.webkit.org/278327@main