LockFairnessTest in ToyLocks benchmark should have a low priority thread. https://bugs.webkit.org/show_bug.cgi?id=273713 rdar://127513197
Reviewed by NOBODY (OOPS!).

With this change our "eventually fair" locks don't seem so fair:

% ./LockFairnessTest lock 50 20 50
WTFLock: 4107, 3932, 4233, 4235, 3732, 4258, 4033, 4086, 4210, 3965, 3979, 4020, 4020, 4090, 4254, 3854, 3824, 4177, 4116, 4208, 4111, 4085, 4298, 4005, 4101, 4169, 4169, 3872, 4184, 4242, 4165, 4382, 3998, 4218, 4110, 4408, 4423, 3873, 4132, 3956, 4593, 4198, 3895, 4395, 4217, 4024, 4284, 4292, 4015, 567

The reason is that every time WTF::Lock barges it throws away the queuing order of ParkingLot. Since
the low priority thread tends to get scheduled last it will go to the end of the queue most times.

Explanation of why this fixes the bug (OOPS!).

* Source/WTF/benchmarks/LockFairnessTest.cpp:
* Source/WTF/benchmarks/ToyLocks.h: