REGRESSION (iOS 17.4.1?): <video> tag fails to adjust its size according to changes in the parent element https://bugs.webkit.org/show_bug.cgi?id=272078 rdar://125881482
Reviewed by Jer Noble.

When layout changes the size of a remote-hosted video element with a MediaStream
source, the order of calls to `VideoPresentationManagerProxy` and
`VideoPresentationModelContext` isn't always the same. If
`VideoPresentationManagerProxy::createViewWithID` is called before the model's
view has been created, it creates and configures a WebAVPlayerLayerView and
a WKVideoView. It did not set the model's video dimensions, which sometimes
caused WebAVPlayerLayer to be resized incorrectly when layout changed the
size of the video element very quickly.

* Source/WebKit/UIProcess/Cocoa/VideoPresentationManagerProxy.mm:
(WebKit::VideoPresentationManagerProxy::createViewWithID): Call model->setVideoDimensions
with the video's native dimensions.

Canonical link: https://commits.webkit.org/278350@main